requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
distlib==0.3.6
filelock==3.8.2
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
platformdirs==2.6.0
pyTelegramBotAPI==4.8.0
python-decouple==3.7
pytz==2022.7.1
requests==2.28.1
urllib3==1.26.13
virtualenv==20.17.1
yarl==1.8.2
